fix: localize case row conflict handling
