+++ tracecat/cases/rows/router.py
@@ -10,10 +10,9 @@ from starlette.status import (
 )
 
 from tracecat import config
-from tracecat.api.common import raise_known_database_http_exception
 from tracecat.auth.dependencies import WorkspaceUserRouteRole
 from tracecat.authz.controls import require_scope
-from tracecat.cases.rows.common import CASE_ROW_LINK_CONSTRAINT_ERRORS
+from tracecat.cases.rows.exceptions import raise_case_row_link_integrity_error
 from tracecat.cases.rows.schemas import (
     CaseTableRowInsertCreate,
     CaseTableRowLinkCreate,
@@ -78,11 +77,7 @@ async def link_case_row(
     except ValueError as exc:
         raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
     except IntegrityError as exc:
-        await raise_known_database_http_exception(
-            session,
-            exc,
-            constraint_errors=CASE_ROW_LINK_CONSTRAINT_ERRORS,
-        )
+        await raise_case_row_link_integrity_error(session, exc)
 
 
 @router.post("/{case_id}/rows/insert", status_code=HTTP_201_CREATED)
